+++ packages/dnr-converter/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Preserve `$csp` allowlist rules in source-map ruleset metadata for MV3 runtime
+  exception planning.
+
+### Fixed
+
+- Prevent `$csp` allowlist rules from being converted into broad DNR `allow` rules.
+
 ## [1.1.0] - 2026-07-28
 
 ### Changed